some issues fixed and pre-commit
